requirements.txt
fastapi
uvicorn
google-generativeai
pydantic
psycopg2-binary